fix(decorator): :bug: More elegantly handle if api key or measurement id aren't set, and reuse gtag tracker object
Have check for if key or measurement fallback are not set as env variables, if not then print a message to user telling them the issue. If there's an issue return a False success condition so that later tracking functions do not run and we don't keep trying. Also pass the identified gtag tracker through function so we don't define it more than once